# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Issue Tracker API
  description: REST API for a Redmine-like issue tracker. Supports projects, issues, comments, and role-based access control.
  license:
    name: ''
  version: 0.1.0
paths:
...
  /api/issues/{issue_id}:
    get:
      tags:
      - Issues
      summary: Get issue
      description: Returns a single issue by ID. Requires visibility access to the parent project.
      operationId: get_issue
      parameters:
      - name: issue_id
        in: path
        description: Issue ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Issue details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Issue not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
    delete:
      tags:
      - Issues
      summary: Delete issue
      description: Deletes an issue. Only the issue author can perform this action.
      operationId: delete_issue
      parameters:
      - name: issue_id
        in: path
        description: Issue ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Issue deleted
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied (not author)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Issue not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
    patch:
      tags:
      - Issues
      summary: Update issue
      description: Updates an issue's fields. Only supplied fields are changed. Requires write access.
      operationId: update_issue
      parameters:
      - name: issue_id
        in: path
        description: Issue ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIssueRequest'
        required: true
      responses:
        '200':
          description: Issue updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Issue not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
...
  /api/projects/{project_id}/issues:
    get:
      tags:
      - Issues
      summary: List issues
      description: Lists issues for a project with optional status, priority, and assignee filters.
      operationId: list_issues
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: priority
        in: query
        required: false
        schema:
          type: string
      - name: assignee_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Paginated list of issues
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedIssueResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
    post:
      tags:
      - Issues
      summary: Create issue
      description: Creates a new issue in a project. Requires write access (owner, admin, or member role).
      operationId: create_issue
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
        required: true
      responses:
        '201':
          description: Issue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
...
components:
  schemas:
...
    CreateIssueRequest:
      type: object
      description: Request body for `POST /api/projects/:project_id/issues`.
      required:
      - subject
      properties:
        assignee_id:
          type:
          - string
          - 'null'
          format: uuid
        description:
          type: string
          default: ''
        due_date:
          type:
          - string
          - 'null'
          format: date
        issue_type:
          type: string
          default: task
        priority:
          type: string
          default: normal
        status:
          type: string
          default: open
        subject:
          type: string
...
    ErrorResponse:
      type: object
      description: |-
        Schema-only type for documenting the standard error response shape in OpenAPI.
        Not used at runtime — the actual JSON is built inline in `AppError::into_response`.
      required:
      - message
      properties:
        data:
          description: Additional error details (validation field errors, etc.) or null.
        message:
          type: string
          description: Human-readable error message.
    Issue:
      type: object
      required:
      - id
      - project_id
      - author_id
      - subject
      - description
      - issue_type
      - status
      - priority
      - created_at
      - updated_at
      properties:
        assignee_id:
          type:
          - string
          - 'null'
          format: uuid
        author_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        description:
          type: string
        due_date:
          type:
          - string
          - 'null'
          format: date
        id:
          type: string
          format: uuid
        issue_type:
          type: string
        priority:
          type: string
        project_id:
          type: string
          format: uuid
        status:
          type: string
        subject:
          type: string
        updated_at:
          type: string
          format: date-time
...
    PaginatedIssueResponse:
      type: object
      required:
      - data
      - total
      - page
      - limit
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Issue'
        limit:
          type: integer
          format: int64
        page:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
...
    UpdateIssueRequest:
      type: object
      description: |-
        Request body for `PATCH /api/issues/:issue_id`.

        All fields are optional — only provided fields are updated.
      properties:
        assignee_id:
          type:
          - string
          - 'null'
          format: uuid
        description:
          type:
          - string
          - 'null'
        due_date:
          type:
          - string
          - 'null'
          format: date
        issue_type:
          type:
          - string
          - 'null'
        priority:
          type:
          - string
          - 'null'
        status:
          type:
          - string
          - 'null'
        subject:
          type:
          - string
          - 'null'
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...